requirements.txt
# Core dependencies for LACRM Sync Tool
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
dnspython>=2.2.0
python-whois>=0.8.0
wappalyzer-py>=0.1.0
openai>=1.0.0
psycopg2-binary>=2.9.0
tqdm>=4.64.0
configparser>=5.3.0
